blacking

+++ deep/feature_extractor.py
@@ -5,11 +5,12 @@ import cv2
 
 from .model import Net
 
+
 class Extractor(object):
     def __init__(self, model_path, use_cuda=True):
         self.net = Net(reid=True)
         self.device = "cuda" if torch.cuda.is_available() and use_cuda else "cpu"
-        state_dict = torch.load(model_path)['net_dict']
+        state_dict = torch.load(model_path)["net_dict"]
         self.net.load_state_dict(state_dict)
         print("Loading weights from {}... Done!".format(model_path))
         self.net.to(self.device)
@@ -17,9 +18,9 @@ class Extractor(object):
 
     def __call__(self, img):
         assert isinstance(img, np.ndarray), "type error"
-        img = img.astype(np.float)#/255.
-        img = cv2.resize(img, (64,128))
-        img = torch.from_numpy(img).float().permute(2,0,1)
+        img = img.astype(np.float)  # /255.
+        img = cv2.resize(img, (64, 128))
+        img = torch.from_numpy(img).float().permute(2, 0, 1)
         img = self.norm(img).unsqueeze(0)
         with torch.no_grad():
             img = img.to(self.device)
@@ -27,9 +28,8 @@ class Extractor(object):
         return feature.cpu().numpy()
 
 
-if __name__ == '__main__':
-    img = cv2.imread("demo.jpg")[:,:,(2,1,0)]
+if __name__ == "__main__":
+    img = cv2.imread("demo.jpg")[:, :, (2, 1, 0)]
     extr = Extractor("checkpoint/ckpt.t7")
     feature = extr(img)
     print(feature.shape)
-
